Make code style consistent, tsify, clean up
