--- .npmignore
@@ -1,4 +1,0 @@
-# meghatározza hogy mi ne legyen benne az npm package-ben
-src
-example
-*.py
