--- .npmrc
@@ -1,1 +1,0 @@
-@Kris030:registry=https://npm.pkg.github.com
