--- example/forceStop.js
@@ -1,8 +1,0 @@
-import { Robot } from '../out/lib_class.js';
-
-const roland = new Robot();
-
-await roland.init();
-console.log('gasdfgsdgf');
-
-roland.move();
