--- example/handRecognition/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <script src="https://cdn.jsdelivr.net/npm/@tensorflow/tfjs@1.3.1/dist/tf.min.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/@teachablemachine/image@0.8.3/dist/teachablemachine-image.min.js"></script>
-        
-        <script src="https://cdn.socket.io/3.1.3/socket.io.min.js" integrity="sha384-cPwlPLvBTa3sKAgddT6krw0cJat7egBga3DJepJyrLl4Q9/5WLra3rrnMcyTyOnh" crossorigin="anonymous"></script>
-
-        
-    </head>
-    <div>Teachable Machine Image Model</div>
-
-    <button type='button' id="start-button">Start</button>
-    <div id='webcam-container'></div>
-    <div id='label-container'></div>
-    
-    <script src="index.js" type="module"></script>
-</html>
