+++ package.json
@@ -1,30 +1,33 @@
 {
-  "name": "@kris030/roblib",
-  "version": "1.0.0",
-  "description": "A remote library for a dank engine.",
-  "type": "module",
-  "main": "out/lib.js",
-  "dependencies": {
-    "@google-cloud/speech": "^4.5.6",
-    "dotenv": "^10.0.0",
-    "node-fetch": "^2.6.1",
-    "node-record-lpcm16": "^1.0.1",
-    "socket.io-client": "^4.1.3"
-  },
-  "devDependencies": {
-    "typescript": "^4.3.5"
-  },
-  "scripts": {
-    "dev": "node --enable-source-maps --experimental-json-modules --no-warnings .",
-    "build": "tsc",
-    "start": "node --experimental-json-modules .",
-    "clean": "rm -rf out"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kris030/roblib.git"
-  },
-  "publishConfig": {
-    "registry": "https://npm.pkg.github.com"
-  }
+    "name": "@kris030/roblib",
+    "version": "1.0.0",
+    "description": "A remote library for a dank engine.",
+    "type": "module",
+    "main": "out/lib.js",
+    "engines": {
+        "node": ">=14.8"
+    },
+    "dependencies": {
+        "@google-cloud/speech": "^4.9.0",
+        "dotenv": "^10.0.0",
+        "node-fetch": "^2.6.1",
+        "node-record-lpcm16": "^1.0.1",
+        "socket.io-client": "^4.1.3"
+    },
+    "devDependencies": {
+        "typescript": "^4.3.5"
+    },
+    "scripts": {
+        "dev": "node --enable-source-maps --experimental-json-modules --no-warnings .",
+        "build": "tsc",
+        "watch": "tsc -w",
+        "start": "node --experimental-json-modules .",
+        "stop": "node out/forceStop.js",
+        "clean": "rm -rf out",
+        "postinstall": "npx tsc -p ."
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/Kris030/roblib.git"
+    }
 }
