+++ server/start.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+FLASK_APP=server.py python3 -m flask run
