+++ src/example/handRecognition/metadata.json
