--- utils.py
@@ -1,7 +1,0 @@
-
-def clamp(val, min, max):
-	if val < min:
-		return min
-	if val > max:
-		return max
-	return val
